Refactored the entire codes - Split sections into react components
